Remove failure and instability conditions as it's handled by Gatling itself
