+++ src/main/java/com/excilys/ebi/gatling/jenkins/GatlingProjectAction.java
@@ -21,7 +21,6 @@ import hudson.model.AbstractBuild;
 import hudson.model.AbstractProject;
 
 import java.util.LinkedHashMap;
-import java.util.List;
 import java.util.Map;
 
 import com.excilys.ebi.gatling.jenkins.chart.Graph;
@@ -53,7 +52,7 @@ public class GatlingProjectAction implements Action {
 	public boolean isVisible() {
 		for (AbstractBuild<?, ?> build : getProject().getBuilds()) {
 			GatlingBuildAction gatlingBuildAction = build.getAction(GatlingBuildAction.class);
-			if (gatlingBuildAction != null && gatlingBuildAction.getRequestsReports() != null && gatlingBuildAction.getRequestsReports().size() > 0) {
+			if (gatlingBuildAction != null) {
 				return true;
 			}
 		}
@@ -96,21 +95,19 @@ public class GatlingProjectAction implements Action {
 		};
 	}
 
-	public Map<AbstractBuild<?, ?>, List<String>> getReports() {
-		Map<AbstractBuild<?, ?>, List<String>> reports = new LinkedHashMap<AbstractBuild<?, ?>, List<String>>();
+	public Map<AbstractBuild<?, ?>, String> getReports() {
+		Map<AbstractBuild<?, ?>, String> reports = new LinkedHashMap<AbstractBuild<?, ?>, String>();
 
 		for (AbstractBuild<?, ?> build : project.getBuilds()) {
-			List<String> simulations = GatlingBuildAction.getReports(build);
-
-			if (!simulations.isEmpty()) {
-				reports.put(build, simulations);
-			}
+			GatlingBuildAction action = build.getAction(GatlingBuildAction.class);
+			if (action != null)
+				reports.put(build, action.getSimulationName());
 		}
 
 		return reports;
 	}
 
-	public String getReportURL(int build, String simulation) {
-		return new StringBuilder().append(build).append("/").append(GatlingReportAction.getURL(simulation)).toString();
+	public String getReportURL(int build) {
+		return new StringBuilder().append(build).append("/").append(URL_NAME).toString();
 	}
 }
