+++ src/main/resources/com/excilys/ebi/gatling/jenkins/GatlingBuildAction/index.properties
@@ -1,2 +1,1 @@
-BuildPageTitle=Report(s) for build \#
-AvailableReports=Available reports for this build
+OpenNewPage=Open in a new page.
