+++ src/main/resources/com/excilys/ebi/gatling/jenkins/GatlingBuildAction/index_fr.properties
@@ -1,2 +1,1 @@
-BuildPageTitle=Rapport(s) de la build \#
-AvailableReports=Rapports disponibles pour cette build
+OpenNewPage=Ouvrir dans une autre page.
