+++ src/main/resources/com/excilys/ebi/gatling/jenkins/GatlingProjectAction/floatingBox.jelly
@@ -4,7 +4,7 @@
 		<div class="test-trend-caption">
 			${%PerformanceTrend}
 			<g:graph id="dashboardGatling" seriesNames="${action.dashboardGraph.seriesNamesJSON}"
-			         data="${action.dashboardGraph.seriesJSON}" yAxisUnit="ms" height="400" width="200"/>
+			         data="${action.dashboardGraph.seriesJSON}" yAxisUnit="ms" height="400px" width="100%"/>
 		</div>
 	</j:if>
 </j:jelly>
