+++ src/main/resources/com/excilys/ebi/gatling/jenkins/GatlingPublisher/help_fr.html
@@ -1,27 +1,8 @@
-<!--
-
-    Copyright 2011-2012 eBusiness Information, Groupe Excilys (www.excilys.com)
-
-    Licensed under the Apache License, Version 2.0 (the "License");
-    you may not use this file except in compliance with the License.
-    You may obtain a copy of the License at
-
-    		http://www.apache.org/licenses/LICENSE-2.0
-
-    Unless required by applicable law or agreed to in writing, software
-    distributed under the License is distributed on an "AS IS" BASIS,
-    WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-    See the License for the specific language governing permissions and
-    limitations under the License.
-
--->
 <div>
 	<p>
 		Ce plugin permet d'intégrer le suivit d'une simulation Gatling à votre job Jenkins :
 	</p>
 	<ul>
-		<li>Mise en place de conditions d'échec ou d'instablilité du build en fonction des résultats de la simulation
-		</li>
 		<li>Suivi des performances au cours du temps</li>
 		<li>Publication des rapports</li>
 	</ul>
