--- src/main/resources/com/excilys/ebi/gatling/jenkins/GatlingReportAction/index.jelly
@@ -1,23 +1,0 @@
-<j:jelly xmlns:j="jelly:core" xmlns:st="jelly:stapler"
-         xmlns:l="/lib/layout">
-	<l:layout title="Build #${it.build.number} : ${it.simulationName}">
-		<l:breadcrumb title="Gatling" href="../gatling"/>
-		<l:breadcrumb title="${it.simulationName}" href="."/>
-		<l:side-panel>
-			<st:include it="${it.build}" page="sidepanel.jelly"/>
-		</l:side-panel>
-		<l:main-panel>
-			<h1>
-				<a href="source" target="_blank">${%OpenNewPage}</a>
-			</h1>
-			<iframe id="reportFrame" src="source" width="100%" height="100%" frameborder="0"
-			        onload="resizeReportFrame()"></iframe>
-			<script type="text/javascript">
-				function resizeReportFrame() {
-					var frame = document.getElementById("reportFrame")
-					frame.height = frame.contentWindow.document.body.scrollHeight + "px"
-				}
-			</script>
-		</l:main-panel>
-	</l:layout>
-</j:jelly>
