--- src/main/resources/com/excilys/ebi/gatling/jenkins/GatlingReportAction/index.properties
@@ -1,1 +1,0 @@
-OpenNewPage=Open in a new page.
