--- src/main/resources/com/excilys/ebi/gatling/jenkins/GatlingReportAction/index_fr.properties
@@ -1,1 +1,0 @@
-OpenNewPage=Ouvrir dans une nouvelle page.
