--- src/main/resources/com/excilys/ebi/gatling/jenkins/model/Condition/config.jelly
@@ -1,13 +1,0 @@
-<j:jelly xmlns:j="jelly:core"
-         xmlns:f="/lib/form">
-	<f:entry title="${%Property}">
-		<f:enum field="property">${it.description}</f:enum>
-	</f:entry>
-	<f:entry title="${%Threshold}">
-		<f:number field="conditionThreshold"
-		          checkUrl="'descriptorByName/com.excilys.ebi.gatling.jenkins.model.Condition/checkConditionThreshold?val='+this.value"/>
-	</f:entry>
-	<f:entry>
-		<f:repeatableDeleteButton value="${%Delete}"/>
-	</f:entry>
-</j:jelly>
