--- src/main/resources/com/excilys/ebi/gatling/jenkins/model/Condition/config.properties
@@ -1,3 +1,0 @@
-Property=Property
-Threshold=Threshold
-Delete=Delete condition
