--- src/main/resources/com/excilys/ebi/gatling/jenkins/model/Condition/config_fr.properties
@@ -1,3 +1,0 @@
-Property=Propri\u00e9t\u00e9
-Threshold=Seuil
-Delete=Supprimer la condition
