+++ src/main/resources/com/excilys/ebi/gatling/jenkins/model/Simulation/config.jelly
@@ -4,10 +4,4 @@
 		<f:textbox field="name" value="${instance.name}"
 		           checkUrl="'descriptorByName/com.excilys.ebi.gatling.jenkins.model.Simulation/checkName?val='+this.value"/>
 	</f:entry>
-	<f:entry title="${%Fail}">
-		<g:repeatableProperty field="failConditions" add="${%AddFail}"/>
-	</f:entry>
-	<f:entry title="${%Unstable}">
-		<g:repeatableProperty field="unstableConditions" add="${%AddUnstable}"/>
-	</f:entry>
 </j:jelly>
