--- src/main/resources/com/excilys/ebi/gatling/jenkins/tags/repeatableProperty.jelly
@@ -1,47 +1,0 @@
-<?jelly escape-by-default='true'?>
-<j:jelly xmlns:j="jelly:core" xmlns:st="jelly:stapler"
-         xmlns:f="/lib/form">
-	<st:documentation> <![CDATA[]
-    Data-bound only version of <f:repeatable> that assumes the type pointed by the property is data-bound as well.
-    The nested property type must be Describable and it needs to have config.jelly.
-
-    The nested configuration fragment normally needs to have a delete button by adding a fragment like this:
-
-      	<f:entry title="">
-          <div align="right">
-            <f:repeatableDeleteButton />
-          </div>
-        </f:entry>
-    ]]>
-
-		<st:attribute name="field">
-			Used for the data binding.
-		</st:attribute>
-		<st:attribute name="default">
-			The default value to use for this collection when 'instance[field]' is null.
-		</st:attribute>
-		<st:attribute name="noAddButton">
-			true if the default 'add' button (that adds a new copy) shouldn't be displayed.
-			When you use this attribute,
-		</st:attribute>
-		<st:attribute name="add">
-			If specified, this text will replace the standard "Add" text.
-		</st:attribute>
-		<st:attribute name="minimum">
-			At least provide this number of copies initially.
-			minimum="1" is useful to make sure there's always at least one entry for the user to fill in.
-		</st:attribute>
-		<st:attribute name="header">
-			For each item, add this header.
-			This also activates drag&amp;drop (where the header is a grip).
-		</st:attribute>
-	</st:documentation>
-
-
-	<f:repeatable field="${attrs.field}" default="${attrs.default}" noAddButton="${attrs.noAddButton}"
-	              header="${attrs.header}" add="${attrs.add}">
-		<table style="width:100%">
-			<st:include page="config.jelly" class="${descriptor.clazz}"/>
-		</table>
-	</f:repeatable>
-</j:jelly>
